LayoutTests/editing/inserting/insert-list-in-table-cell-08.html

<!DOCTYPE html>
<div id="container" contenteditable="true"><table border="1" id="element"><tr><td>fsdf</td><td>fsdf</td></tr><tr><td>gghfg</td><td>fsfg</td></tr></table></div>
<script src="../editing.js"></script>
<script src="../../resources/dump-as-markup.js"></script>
<script>
    Markup.description('Exec insertOrderedList twice in a table removes the previously inserted list item:');

    var e = document.getElementById("element");
    setSelectionCommand(e, 0, e, 1);
    Markup.dump('container', 'Before');

    document.execCommand("insertOrderedList");
    document.execCommand("insertOrderedList");
    Markup.dump('container', 'After');
</script>
